fix: Fixing some bugs

+++ app/db/migrations/versions/873ef9b26213_initial_db.py
@@ -53,7 +53,7 @@ CREATE TABLE {_env.ENVIRONMENT}.modalities (
 CREATE TABLE {_env.ENVIRONMENT}.properties (
 	id serial4 NOT NULL,
 	company_id int4 NOT NULL,
-	code int4 NOT NULL,
+	code int8 NOT NULL,
 	title varchar(255) NULL,
 	price double precision NOT NULL,
 	description varchar NULL,
